Crate solana_transaction_context

Data shared between program runtime and built-in programs as well as SBF programs.

Modules

transaction_accounts

vm_slice

Structs

BorrowedInstructionAccount

Shared account borrowed from the TransactionContext and an InstructionContext.

ExecutionRecord

Everything that needs to be recorded from a TransactionContext after execution

InstructionAccount

Contains account meta data which varies between instruction.

InstructionContext

View interface to read instructions.

InstructionFrame

Instruction shared between runtime and programs.

TransactionContext

Loaded transaction shared between runtime and programs.

TransactionReturnData

Return data at the end of a transaction

Constants

MAX_ACCOUNTS_PER_INSTRUCTION

MAX_ACCOUNTS_PER_TRANSACTION

MAX_ACCOUNT_DATA_GROWTH_PER_INSTRUCTION

MAX_ACCOUNT_DATA_GROWTH_PER_TRANSACTION

MAX_ACCOUNT_DATA_LEN

MAX_INSTRUCTION_DATA_LEN

MAX_INSTRUCTION_TRACE_LENGTH

Type Aliases

IndexOfAccount

Index of an account inside of the transaction or an instruction.