solana_transaction_context

Struct TransactionContext

pub struct TransactionContext<'ix_data> { /* private fields */ }

Loaded transaction shared between runtime and programs.

This context is valid for the entire duration of a transaction being processed.

Implementations

impl<'ix_data> TransactionContext<'ix_data>

pub fn new( transaction_accounts: Vec<KeyedAccountSharedData>, rent: Rent, instruction_stack_capacity: usize, instruction_trace_capacity: usize, ) -> Self

Constructs a new TransactionContext

pub fn deconstruct_without_keys( self, ) -> Result<Vec<AccountSharedData>, InstructionError>

Used in mock_process_instruction

pub fn accounts(&self) -> &Rc<TransactionAccounts>

pub fn get_number_of_accounts(&self) -> IndexOfAccount

Returns the total number of accounts loaded in this Transaction

pub fn get_key_of_account_at_index( &self, index_in_transaction: IndexOfAccount, ) -> Result<&Pubkey, InstructionError>

Searches for an account by its key

pub fn find_index_of_account(&self, pubkey: &Pubkey) -> Option<IndexOfAccount>

Searches for an account by its key

pub fn get_instruction_trace_capacity(&self) -> usize

Gets the max length of the instruction trace

pub fn get_instruction_trace_length(&self) -> usize

Returns the instruction trace length.

Not counting the last empty instruction which is always pre-reserved for the next instruction. See also get_next_instruction_context().

pub fn get_instruction_context_at_index_in_trace( &self, index_in_trace: usize, ) -> Result<InstructionContext<'_, '_>, InstructionError>

Gets a view on an instruction by its index in the trace

pub fn get_instruction_context_at_nesting_level( &self, nesting_level: usize, ) -> Result<InstructionContext<'_, '_>, InstructionError>

Gets a view on the instruction by its nesting level in the stack

pub fn get_instruction_stack_capacity(&self) -> usize

Gets the max height of the instruction stack

pub fn get_instruction_stack_height(&self) -> usize

Gets instruction stack height, top-level instructions are height solana_instruction::TRANSACTION_LEVEL_STACK_HEIGHT

pub fn get_current_instruction_context( &self, ) -> Result<InstructionContext<'_, '_>, InstructionError>

Returns a view on the current instruction

pub fn get_next_instruction_context( &self, ) -> Result<InstructionContext<'_, '_>, InstructionError>

Returns a view on the next instruction. This function assumes it has already been configured with the correct values in prepare_next_instruction or prepare_next_top_level_instruction

pub fn configure_next_instruction( &mut self, program_index: IndexOfAccount, instruction_accounts: Vec<InstructionAccount>, deduplication_map: Vec<u16>, instruction_data: Cow<'ix_data, [u8]>, ) -> Result<(), InstructionError>

Configures the next instruction.

The last InstructionContext is always empty and pre-reserved for the next instruction.

pub fn configure_next_instruction_for_tests( &mut self, program_index: IndexOfAccount, instruction_accounts: Vec<InstructionAccount>, instruction_data: Vec<u8>, ) -> Result<(), InstructionError>

A version of configure_next_instruction to help creating the deduplication map in tests

pub fn push(&mut self) -> Result<(), InstructionError>

Pushes the next instruction

pub fn pop(&mut self) -> Result<(), InstructionError>

Pops the current instruction

pub fn get_return_data(&self) -> (&Pubkey, &[u8])

Gets the return data of the current instruction or any above

pub fn set_return_data( &mut self, program_id: Pubkey, data: Vec<u8>, ) -> Result<(), InstructionError>

Set the return data of the current instruction

pub fn access_violation_handler( &self, stricter_abi_and_runtime_constraints: bool, account_data_direct_mapping: bool, ) -> AccessViolationHandler

Returns a new account data write access handler

pub fn take_instruction_trace(&mut self) -> Vec<InstructionFrame<'_>>

Take ownership of the instruction trace